// src/information/statics.d.ts
declare const icons: {
    email: string;
    linkedin: string;
    github: string;
    discord: string;
    instagram: string;
    copyright: string;
    resume: string;
    check_mark: string;
};
declare const logos: {
    instagram_logo: string;
    discord_logo: string;
    linkedin_logo: string;
    java: string;
    python: string;
    javascript: string;
    typescript: string;
    cplus: string;
    csharp: string;
    html: string;
    css: string;
    sql: string;
    react: string;
    nodejs: string;
    git: string;
    mongodb: string;
    flask: string;
    socketio: string;
    tailwind: string;
    expo: string;
};
declare const thumbnails: {
    portfolio_tn: string;
    todo_tn: string;
    live_chat_tn: string;
};
export { icons, logos, thumbnails };